added minimal sort plate
modified the sort plate to what I believe to be the minimum viable solution.  I removed all of the sort pipe holes and added indexing notches so that position calibration can still be done. This is for use with the single sort pipe that was show in the last update video.